Code Performance
There are three files added to the repository jroMI;

covXYs_Measured.py : time measurements of the modules is done.
Profiling is used to see the process time of each commands in the
modules.

covXY_Performance_Results.rtf : the results of performance measurement
is reported.

covXY_Grid_2D_modified.ipynb : Data is generated by "geom" module in a
seperate loop and the results of this is shown. Besides, h5py comes to
the scene. It`s fun to play with h5py.