Some little change in MIprofile.py and - factor in the formula of CovXYs.py
It was not showing the last subfigure. Fixed.
Also I made the subfigues look nicer (sharing the same axis, etc.)
In CovXYs.py, I notiiced that even thhough exponential term is a factor
of "-"  in Maro Milla's thesis and Z.Feng`s paper, it was "+" in the
slides given to me before. Hence, I swicthed the sign from '+' to '-'.